Try to make the linter happy
