+++ .dockerignore
@@ -14,4 +14,5 @@ mainapp/assets/bundles
 .git
 .idea
 .tox
-.env
+.env*
+*.log
