+++ mainapp/models/user_alert.py
@@ -5,7 +5,10 @@ from mainapp.functions.search_tools import (
     params_to_search_string,
     search_string_to_params,
 )
-from mainapp.functions.search_notification_tools import params_to_human_string, params_are_equal
+from mainapp.functions.search_notification_tools import (
+    params_to_human_string,
+    params_are_equal,
+)
 
 
 class UserAlert(models.Model):
