+++ mainapp/tests/test_rss.py
@@ -11,7 +11,7 @@ class TestRSS(TestCase):
     c = Client()
 
     def test_paper_feed(self):
-        response = self.c.get("/paper/feed/").content.decode().strip()
+        response = self.c.get("/paper/feed/").content.decode()
         self.assertIn("<rss", response)
         self.assertIn("Latest papers", response)
         self.assertIn(
@@ -24,7 +24,7 @@ class TestRSS(TestCase):
         new=MockMainappSearch.execute,
     )
     def test_search_results(self):
-        response = self.c.get("/search/query/complexity/feed/").content.decode().strip()
+        response = self.c.get("/search/query/complexity/feed/").content.decode()
         self.assertIn("<rss", response)
         self.assertIn("The latest search results", response)
         self.assertIn("File: Title", response)
