+++ mainapp/views/__init__.py
@@ -3,3 +3,4 @@ from .meeting import *
 from .persons import *
 from .search import *
 from .views import *
+from .special import *
